Fix bug when using the profitable_listings view. All five total_profit expressions now explicitly cast to ::bigint. The root cause is a PostgreSQL quirk worth   knowing: SUM(int2) and SUM(int4) return bigint, but SUM(int8) returns numeric — because summing 64-bit integers can   overflow 64-bit, so PostgreSQL silently promotes. With ::bigint on the outer SUM, PostgreSQL returns int8 and Npgsql   uses its fast integer converter rather than going through the decimal path.

+++ .docker/flyway/sql/Repeatable/R__view_profitable_listings_view.sql
@@ -21,7 +21,7 @@ WITH
       MIN(l.price)                                              AS min_price,
       MAX(l.price)                                              AS max_price,
       SUM(l.quantity)::int                                      AS quantity,
-      SUM((i.value_sell_price - l.price) * l.quantity)         AS total_profit
+      SUM((i.value_sell_price - l.price) * l.quantity)::bigint  AS total_profit
     FROM public.listings l
     JOIN public.items i ON i.id = l.item_id
     WHERE l.source = 'Torn'
@@ -36,7 +36,7 @@ WITH
       MIN(l.price)                                              AS min_price,
       MAX(l.price)                                              AS max_price,
       SUM(l.quantity)::int                                      AS quantity,
-      SUM((wb.min_price - l.price) * l.quantity)               AS total_profit
+      SUM((wb.min_price - l.price) * l.quantity)::bigint        AS total_profit
     FROM public.listings l
     JOIN weav3r_base wb ON wb.item_id = l.item_id
     WHERE l.source = 'Torn'
@@ -50,7 +50,7 @@ WITH
       MIN(l.price)                                              AS min_price,
       MAX(l.price)                                              AS max_price,
       SUM(l.quantity)::int                                      AS quantity,
-      SUM((i.value_sell_price - l.price) * l.quantity)         AS total_profit
+      SUM((i.value_sell_price - l.price) * l.quantity)::bigint  AS total_profit
     FROM public.listings l
     JOIN public.items i ON i.id = l.item_id
     WHERE l.source = 'Weav3r'
@@ -65,7 +65,7 @@ WITH
       MIN(l.price)                                                               AS min_price,
       MAX(l.price)                                                               AS max_price,
       SUM(l.quantity)::int                                                       AS quantity,
-      SUM((FLOOR(i.value_market_price * 0.95)::bigint - l.price) * l.quantity) AS total_profit
+      SUM((FLOOR(i.value_market_price * 0.95)::bigint - l.price) * l.quantity)::bigint AS total_profit
     FROM public.listings l
     JOIN public.items i ON i.id = l.item_id
     WHERE l.source = 'Weav3r'
@@ -80,7 +80,7 @@ WITH
       MIN(l.price)                                                               AS min_price,
       MAX(l.price)                                                               AS max_price,
       SUM(l.quantity)::int                                                       AS quantity,
-      SUM((FLOOR(i.value_market_price * 0.85)::bigint - l.price) * l.quantity) AS total_profit
+      SUM((FLOOR(i.value_market_price * 0.85)::bigint - l.price) * l.quantity)::bigint AS total_profit
     FROM public.listings l
     JOIN public.items i ON i.id = l.item_id
     WHERE l.source = 'Weav3r'
